fix: 将s2s过期时间从i32修改为i64

+++ s2sname/s2sname.thrift
@@ -7,7 +7,7 @@ struct s2sname {
     2:i32   port,
     3:i32   prority,
     4:string name,
-    5:i32   Expired;
+    5:i64   Expired;
 }
 
 struct register_req {
